requirements.txt
python-telegram-bot==13.8
environs==14.3.0
urllib3==1.26.20